move getHosname to PosixSystem
